Support a Python lambda inside the replacement \G{<code>}
These specifiers are interpolated into a Python eval under a
namespace of useful functions. For example, 'G' is the match object
and 'm(i)' is shorthand for 'G.groups(0)[i]'. There could be
all sorts of string manipulation functions in the namespace.

For backward compatiblity, only if '\G{' is present in the replacing
string, this feature is activated. When it is active, the standard
group specifiers, e.g. \1 no longer work.

For example:

    gsr '(GitS[a-z]+Replace)///\G{titlecase_to_underscore(m(0))}' --diff

gitsearchreplace/__init__.py
--- a/gitsearchreplace/__init__.py 17:35:02.284896013 +0300
+++ b/gitsearchreplace/__init__.py 17:35:49.412636678 +0300
@@ -37,7 +37,7 @@
 def titlecase_to_underscore(name):
     return re.sub('([a-z0-9])([A-Z])', r'\1_\2', name).lower()

-class GitSearchReplace(object):
+class git_search_replace(object):
     """Main class"""

     def __init__(self, separator=None, diff=None, fix=None, exclude=None, expressions=None):
@@ -239,7 +239,7 @@
                       "pattern (can be specified more than once)")

     (options, args) = parser.parse_args()
-    gsr = GitSearchReplace(
+    gsr = git_search_replace(
         separator=options.separator,
         diff=options.diff,
         fix=options.fix,